+++ custom_addons/ak_library_management/controllers/main.py
@@ -3,6 +3,7 @@
 from odoo import http
 from odoo.http import request
 
+
 class ContactsController(http.Controller):
 
     @http.route('/contacts', type="http", auth="public", website=True)
@@ -12,3 +13,11 @@ class ContactsController(http.Controller):
             'records': contacts
         }
         return request.render('ak_library_management.contact_kanban_template', values)
+
+    @http.route('/contacts/<slug>', type="http", auth="public", website=True)
+    def individual_contact(self, slug):
+        contact = request.env['res.partner'].search([('contact_slug', '=', slug)])
+        values = {
+            'contact': contact
+        }
+        return request.render('ak_library_management.contact_detail_form_template', values)